move the rest of queries
