+++ src/lib/sanity.queries/hero/types.ts
@@ -1,0 +1,11 @@
+import type { ImageAsset } from '@sanity/types'
+
+export interface HeroType {
+  backgroundImage: ImageAsset
+  title: string
+  description: string
+  buttonName: {
+    buttonText: string
+    redirectTo: string
+  }
+}
